requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.25
aiosqlite==0.19.0
google-cloud-tasks==2.14.1
